fix: resolve 28 maintainability violations
Python:
- Remove dead legacy verify helpers, update tests
- Move math import to top-level in scoring internals
- Add env parameter to get_action_api_port, get_dashboard_port, _validate_url
- Replace mutable global cache with lru_cache in filesystem helpers
- Add port comment in launcher, rename menubar path constant

UI:
- Extract buildRunSummary, EvalCards to shared modules
- Import FileCopyBtn instead of duplicating in LiveViolationsFeed
- Refactor nested ternaries in TrendBadge to if/else
- Fix variable shadowing and duplicate splitScore in RunOverviewPanel
- Remove useless template literal in ProjectsPage
- Extract magic numbers to named constants
- Add debug logging to empty catch blocks
- Split explorerUtils tests into two files
- Use MODEL_STORAGE_PREFIX consistently in SettingsPage
