+++ src/quodeq/analysis/subagents/verify.py
@@ -199,45 +199,3 @@ def load_previous_findings_for_dimension(
         )
     _findings_cache[cache_key] = (surviving, len(prev_findings), gone)
     return surviving
-
-
-# ---------------------------------------------------------------------------
-# Legacy helpers — kept for backward compatibility with tests
-# ---------------------------------------------------------------------------
-
-def _mechanical_check(finding: dict, src: Path) -> str:
-    """Check if a finding's file still exists. Used by pre-filter."""
-    rel_path = finding.get("file", "")
-    if not rel_path:
-        return "gone"
-    if not (src / rel_path).exists():
-        return "gone"
-    return "ambiguous"  # needs AI verification
-
-
-def run_mechanical_verify(
-    src: Path,
-    prev_findings: list[dict],
-) -> tuple[list[dict], int, list[dict]]:
-    """Pre-filter findings by file existence.
-
-    Returns (surviving, gone_count, []) — all surviving findings
-    are treated as needing AI verification (ambiguous).
-    """
-    surviving, gone = _pre_filter_gone(prev_findings, src)
-    return [], gone, surviving  # none confirmed mechanically, all ambiguous
-
-
-def _write_finding(finding: dict, output_fh: Any) -> None:
-    """Append a finding to the JSONL output file."""
-    output_fh.write(json.dumps(finding) + "\n")
-    output_fh.flush()
-
-
-def write_verified_findings(findings: list[dict], output_jsonl: Path) -> None:
-    """Write verified findings to the JSONL evidence file."""
-    if not findings:
-        return
-    with open(output_jsonl, "a") as fh:
-        for finding in findings:
-            _write_finding(finding, fh)
