+++ src/quodeq/ui/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "test": "node --test src/explorerUtils.test.js",
+    "test": "node --test 'src/**/*.test.js'",
     "build": "vite build",
     "preview": "vite preview"
   },
