+++ src/quodeq/ui/src/features/explorer/components/PrincipleDetailPage.jsx
@@ -65,13 +65,16 @@ function buildViolationPlanText(v, principleName) {
   return lines.join('\n').trim();
 }
 
+const ANIM_DELAY_PER_ITEM_MS = 30;
+const ANIM_MAX_DELAY_MS = 300;
+
 function ViolationCard({ v, principleName, index }) {
   const { filePath, line } = parseFileRef(v.file, v.line);
   const filename = filePath ? filePath.split('/').pop() : null;
   const ref = line != null ? `${filePath}:${line}` : filePath;
   const display = line != null ? `${filename}:${line}` : filename;
   return (
-    <div className={`vdetail-row vdetail-row--${v.severity}`} style={{ animationDelay: `${Math.min(index * 30, 300)}ms` }}>
+    <div className={`vdetail-row vdetail-row--${v.severity}`} style={{ animationDelay: `${Math.min(index * ANIM_DELAY_PER_ITEM_MS, ANIM_MAX_DELAY_MS)}ms` }}>
       <div className="vdetail-row-main">
         <span className={`severity-tag ${v.severity}`}>{v.severity}</span>
         <span className="vrow-label">[{principleName}]</span>
